Fold the hydrate skeleton into the base and drop dead seams
